requirements.txt
aiogram==3.4.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
transaction==4.0
python-dotenv==1.0.1
redis==4.5.4
zope.sqlalchemy==3.1
transformers==4.38.2
torch==2.2.1
sentencepiece==0.2.0
